Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
pillow>=10.0.0
numpy>=1.24.0
python-multipart==0.0.6
pydantic==2.5.0
reportlab==4.0.9
python-dotenv==1.0.0
